requirements.txt
streamlit==1.33.0
qdrant-client==1.9.0
fastembed==0.3.2
ollama==0.1.9
pillow==10.3.0
opencv-python==4.9.0.80
numpy==1.26.4
pandas==2.2.2
tqdm==4.66.4
python-dotenv==1.0.1
insightface
